server/pyproject.toml
[project]
name = "savantsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "beautifulsoup4>=4.13.4",
    "certifi==2025.8.3",
    "click==8.2.1",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "openai==1.99.4",
    "pybaseball>=2.2.7",
    "pydantic==2.11.7",
    "pydantic-ai>=0.7.0",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==14.1.0",
    "rich-toolkit==0.14.9",
    "rignore==0.6.4",
    "sentry-sdk==2.34.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.47.2",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]
